Created the restore command
