+++ src/commands/add.rs
@@ -1,4 +1,5 @@
-use crate::file_ops::{is_url_present, append_to_file};
+use crate::file_ops::is_url_present;
+use crate::file_ops::write_file::append_to_file;
 use std::path::PathBuf;
 use crate::scraper::find_last_chapter;
 
@@ -7,7 +8,7 @@ use crate::scraper::find_last_chapter;
 /// If the CSV file is not present,  another error message is shown.
 /// The function searches for the last chapter at the time, and adds it to the CSV with the URL.
 /// # Arguments:
-/// * `path`: the optional path to the CSV file. If None, the default path will be used (See [file_ops::extract_path_or_default])
+/// * `path`: the optional path to the CSV file. If None, the default path will be used (See [`crate::file_ops::extract_path_or_default`])
 /// * `manga_url`: The Manganelo URL to the manga page.
 pub async fn add_new_manga(path: Option<PathBuf>,  manga_url: &str, verbose: bool) {
     match is_url_present(path.clone(),manga_url) {
