+++ src/commands/export.rs
@@ -1,9 +1,9 @@
 use std::path::PathBuf;
-use crate::file_ops::export_file;
+use crate::file_ops::write_file::export_file;
 
 /// Exports the CSV file to a folder given in parameter.
 /// # Arguments:
-/// * `original_path`: The path to the source file. If None, the default path will be used (See [file_ops::extract_path_or_default])
+/// * `original_path`: The path to the source file. If None, the default path will be used (See [`crate::file_ops::extract_path_or_default`])
 /// * `to`: The folder in which the CSV file will be copied. If not present, an error message will be displayed.
 pub fn export_data(original_path: Option<PathBuf>, to: Option<PathBuf> ) {
     match to {
