+++ src/file_ops/save.rs
@@ -1,5 +1,4 @@
-
-use crate::file_ops::{extract_path_or_default};
+use crate::file_ops::{extract_path_or_default, extract_restore_path_or_default};
 use std::fs;
 use std::io;
 use std::io::{Error, ErrorKind};
@@ -37,18 +36,21 @@ pub fn backup_file(origin_path: Option<PathBuf>) -> Result<(), io::Error> {
 /// # Errors:
 /// * `io::Unsupported` if the path isn't a file. Will be changed to ErrorKind::IsADirectory in the future.
 /// * `io::InvalidInput` if the path doesn't end with .csv.bak or if an error happened while converting the path to str.
-pub fn restore_file(restore_path: Option<PathBuf>) -> Result<(), io::Error> {
-    let path = extract_path_or_default(&restore_path);
-    if !path.is_file() {
-        return Err(Error::new(ErrorKind::Unsupported, "The path should point to the backup CSV file"))
-    }
+pub fn restore_file(restore_path: &Option<PathBuf>, verbose: &bool) -> Result<(), io::Error> {
+    let path = extract_restore_path_or_default(&restore_path);
     match path.clone().to_str() {
         Some(p) => {
             if !p.ends_with(".csv.bak") {
                 return Err(Error::new(ErrorKind::InvalidInput, "The supplied path is incorrect. The extensions should be .csv.bak."))
             }
+            if !path.is_file() {
+                return Err(Error::new(ErrorKind::Unsupported, "The path should point to the backup CSV file"))
+            }
             let copy_path = &p[0..p.len() - 4];
             fs::copy(path, copy_path)?;
+            if *verbose {
+                println!("Restored CSV from {}", p);
+            }
             Ok(())
         },
         None => Err(Error::new(ErrorKind::InvalidInput, "An error happened while converting the path to String."))
@@ -59,7 +61,8 @@ pub fn restore_file(restore_path: Option<PathBuf>) -> Result<(), io::Error> {
 mod tests {
     use super::*;
     use serial_test::serial;
-    use crate::file_ops::{create_file, update_csv};
+    use crate::file_ops::create_file;
+    use crate::file_ops::write_file::update_csv;
     use crate::models::CSVLine;
 
     #[test]
@@ -102,7 +105,7 @@ mod tests {
         assert!(!path.exists());
         let backup_path = PathBuf::from("mangas.csv.bak");
         assert!(backup_path.exists());
-        restore_file(Some(backup_path))?;
+        restore_file(&Some(backup_path), &true)?;
         // Verify that the CSV has been restored
         assert!(path.exists());
         let restored_content = fs::read_to_string(path.clone())?;
